only regenerate the hmm every time (to satisfy the API, which needs to be fixed), but not the alignment

+++ idepi/util/__init__.py
@@ -318,19 +318,22 @@ def generate_alignment(seqrecords, sto_filename, ref_id_func, opts, load=True):
         # we're assuming pre-aligned because they're all generated from the same refseq
         with open(sto_filename, 'w') as fh:
             SeqIO.write(seqrecords, fh, 'stockholm')
-    else:  # not exists(sto_filename):
+    else:
         try:
             tmphmm = generate_hmm_(opts)
             with open(tmphmm, 'rb') as hmm_fh:
                 hmm = hmm_fh.read()
-            tmpaln = generate_alignment_(seqrecords, tmphmm, opts, refseq=opts.REFSEQ)
-            copyfile(tmpaln, sto_filename)
-            log.debug('finished alignment, output moved to {0:s}'.format(sto_filename))
+            if not exists(sto_filename):
+                try:
+                    tmpaln = generate_alignment_(seqrecords, tmphmm, opts, refseq=opts.REFSEQ)
+                    copyfile(tmpaln, sto_filename)
+                    log.debug('finished alignment, output moved to {0:s}'.format(sto_filename))
+                finally:
+                    if exists(tmpaln):
+                        remove(tmpaln)
         finally:
             if exists(tmphmm):
                 remove(tmphmm)
-            if exists(tmpaln):
-                remove(tmpaln)
 
     if load:
         with open(sto_filename) as fh:
